Winners/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
pymongo[srv]==4.5.0
requests==2.28.2
phidata==2.3.0
certifi==2023.7.22
Pillow>=10.0.0
google-cloud-vision>=3.4.4
dnspython==2.4.2
